[WIP] Some perf improvements + time prints for snapshots

+++ android/android-emugl/host/libs/libOpenglRender/FrameBuffer.cpp
@@ -1483,8 +1483,8 @@ bool FrameBuffer::onLoad(Stream* stream) {
             s_egl.eglLoadAllImages(m_eglDisplay, stream);
         }
 #ifdef SNAPSHOT_PROFILE
-        printf("Texture load time: %ld ms\n",
-                    (android::base::System::get()->getUnixTimeUs()
+        printf("Texture load time: %lld ms\n",
+                    (long long)(android::base::System::get()->getUnixTimeUs()
                     - imgStartTime) / 1000);
 #endif
     }
